+++ CHANGELOG.md
@@ -6,14 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 ### Added
-* option to enable and disable verifier (`active` flag): 
+- option to enable and disable verifier (`active` flag): 
   this will affect only voting process, disabled verifier can still withdraw etc.
+- integration with `ContractRegistry` interface. Contract is now updatable.
 
 ### Changed
-* checking if verifier is created is now done by checking if `id` is not empty
+- checking if verifier is created is now done by checking if `id` is not empty
 
 ### Removed
-* `created` field was removed from verifier structure
+- `created` field was removed from verifier structure
 
 ## [0.2.0] - 2018-12-12
 ### Added